버그 수정
 - 롱노트 틱을 추가하는 로직 수정. (소수점 오차 버그 수정)

 - 롱노트 이미지가 끊겨 보이는 현상 수정.

 - 마디 선이 노트보다 밑으로 가도록 수정.

+++ D-BMS/Assets/Scripts/BMSParser.cs
@@ -350,7 +350,7 @@ public class BMSParser : MonoBehaviour
             currentCount++;
             currentBeat = currentCount / beat16Count * 4.0d;
 
-            if (currentBeat <= prevBeat) { continue; }
+            if ((float)currentBeat <= (float)prevBeat) { continue; }
 
             if (currentCount >= beat16Count)
             {
